#3,5 - SubCommit to backup everything before working on shaders
Working on 2nd batch of tasks, added a bunch of new features and deleted the old easter egg, this is 3-4 days of work, just need a quick backup since i will work on an area that's not of my expertise. Following commit will include more structred description of the work.