#4,5: GunStats
+Previously described stat bonus function works across all weapon parts.

*DEFINITION*

The weapon part system takes inspiration from the famous 2016 meme: PPAP by PIKOTARO (see: https://youtu.be/Ct6BUPvE2sM?si=cF8XVXz_d1buVy4z) There are;
+two tier1 parts: Apple & Pen.
+three tier2 parts: ApplePen & LongPen / PenPen & Pineapple.
+one tier3 part: PineapplePen.
+lastly, one tier4 part: Pen Pineapple Apple Pen (PPAP).

Goal is to acquire the PPAP through merging all the parts, and tier levels are determined by how many t1 items are needed to craft the part in question. At the moment, because there are only three debug fight stages, it's impossible to collect enough items to craft PPAP, so in the weaponMerge scene, there exist debug buttons to spawn and test the necessary items. Tier1 item bonuses start from 5%, Tier2 goes up to 10%..... with PPAP giving a %500 buff, indicating the time for a weapon evaluation has come. I'm planning to make it so that finishing a run with PPAP equipped will create a button to get to the next Gun Evolution, which will be implemented on the next commit.